refactor(viztooltip): move chart.tooltip into viztooltip and fix imports and typings
