--- packages/grafana-ui/src/components/Chart/index.tsx
@@ -1,7 +1,0 @@
-import { Tooltip } from './Tooltip';
-
-const Chart = {
-  Tooltip,
-};
-
-export default Chart;
