+++ packages/grafana-ui/src/components/Graph/Graph.story.tsx
@@ -1,11 +1,9 @@
 import React from 'react';
 import { Graph } from '@grafana/ui';
-import Chart from '../Chart';
 import { dateTime, ArrayVector, FieldType, GraphSeriesXY, FieldColorModeId } from '@grafana/data';
 import { Story } from '@storybook/react';
 import { withCenteredStory } from '../../utils/storybook/withCenteredStory';
-import { TooltipContentProps } from '../Chart/Tooltip';
-import { TooltipDisplayMode } from '../VizTooltip/VizTooltip';
+import { VizTooltip, TooltipDisplayMode, VizTooltipContentProps } from '../VizTooltip';
 import { JSONFormatter } from '../JSONFormatter/JSONFormatter';
 import { GraphProps } from './Graph';
 
@@ -117,12 +115,12 @@ export default {
 export const WithTooltip: Story<GraphProps & { tooltipMode: TooltipDisplayMode }> = ({ tooltipMode, ...args }) => {
   return (
     <Graph {...args}>
-      <Chart.Tooltip mode={tooltipMode} />
+      <VizTooltip mode={tooltipMode} />
     </Graph>
   );
 };
 
-const CustomGraphTooltip = ({ activeDimensions }: TooltipContentProps) => {
+const CustomGraphTooltip = ({ activeDimensions }: VizTooltipContentProps) => {
   return (
     <div style={{ height: '200px' }}>
       <div>Showing currently active active dimensions:</div>
@@ -137,7 +135,7 @@ export const WithCustomTooltip: Story<GraphProps & { tooltipMode: TooltipDisplay
 }) => {
   return (
     <Graph {...args}>
-      <Chart.Tooltip mode={tooltipMode} tooltipComponent={CustomGraphTooltip} />
+      <VizTooltip mode={tooltipMode} tooltipComponent={CustomGraphTooltip} />
     </Graph>
   );
 };
