+++ packages/grafana-ui/src/components/Graph/Graph.test.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { mount } from 'enzyme';
 import Graph from './Graph';
-import Chart from '../Chart';
+import { VizTooltip, TooltipDisplayMode } from '../VizTooltip';
 import { GraphSeriesXY, FieldType, ArrayVector, dateTime, FieldColorModeId } from '@grafana/data';
 
 const series: GraphSeriesXY[] = [
@@ -92,7 +92,7 @@ describe('Graph', () => {
       it("doesn't render tooltip when not hovering over a datapoint", () => {
         const graphWithTooltip = (
           <Graph {...mockGraphProps()}>
-            <Chart.Tooltip mode="single" />
+            <VizTooltip mode={TooltipDisplayMode.Single} />
           </Graph>
         );
 
@@ -105,7 +105,7 @@ describe('Graph', () => {
         // Given
         const graphWithTooltip = (
           <Graph {...mockGraphProps()}>
-            <Chart.Tooltip mode="single" />
+            <VizTooltip mode={TooltipDisplayMode.Single} />
           </Graph>
         );
         const container = mount(graphWithTooltip);
@@ -145,7 +145,7 @@ describe('Graph', () => {
         // Given
         const graphWithTooltip = (
           <Graph {...mockGraphProps(true)}>
-            <Chart.Tooltip mode="multi" />
+            <VizTooltip mode={TooltipDisplayMode.Multi} />
           </Graph>
         );
         const container = mount(graphWithTooltip);
