+++ packages/grafana-ui/src/components/VizTooltip/index.tsx
@@ -1,0 +1,4 @@
+export { VizTooltip, VizTooltipContentProps, VizTooltipProps } from './VizTooltip';
+export { VizTooltipContainer, VizTooltipContainerProps } from './VizTooltipContainer';
+export { SeriesTable, SeriesTableProps, SeriesTableRowProps } from './SeriesTable';
+export { TooltipDisplayMode, VizTooltipOptions } from './models.gen';
